# backend-go/internal/transport/http/router/response_alias.go
package router

import "stock-predict-go/internal/transport/http/response"

// APIResponse 是 response.APIResponse 的类型别名，便于 router 包内直接使用统一响应结构。
type APIResponse = response.APIResponse