# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 股票/基金行情平台 API
  description: 股票与基金行情平台后端服务接口文档，提供基金搜索、股票搜索和市场行情等功能；旧预测路径仅作为已废弃兼容端点保留。错误码体系参见 ErrorCode schema。
  version: 1.0.0
servers:
  - url: http://localhost:5070
    description: 本地开发服务器
...
security: []
paths:
...
  /api/v1/stocks/search:
    get:
      tags: [股票]
      summary: 股票搜索
      description: 根据条件搜索股票列表，支持关键词、行业、市场等多维度筛选
      operationId: searchStocks
      parameters:
        - name: keyword
          in: query
          required: false
          description: 搜索关键词（股票代码或名称）
          schema:
            type: string
        - name: industry
          in: query
          required: false
          description: 行业筛选
          schema:
            type: string
        - name: market
          in: query
          required: false
          description: 市场筛选（如 SH、SZ）
          schema:
            type: string
        - name: sort_by
          in: query
          required: false
          description: 排序字段
          schema:
            type: string
        - name: sort_order
          in: query
          required: false
          description: 排序方向（asc/desc）
          schema:
            type: string
            enum: [asc, desc]
        - name: page
          in: query
          required: false
          description: 页码
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          required: false
          description: 每页数量，最大100
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 股票搜索结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/StockSearchData"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/stock/{stockCode}/detail:
    get:
      tags: [股票]
      summary: 股票详情
      description: 获取股票详细信息，包括基本信息、行情、K线、资金流向、财务数据、股东信息等
      operationId: stockDetail
      parameters:
        - name: stockCode
          in: path
          required: true
          description: 股票代码，6位数字
          schema:
            type: string
            pattern: "^[0-9]{6}$"
      responses:
        "200":
          description: 股票详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/StockDetailData"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/stock/{stockCode}/minute:
    get:
      tags: [股票]
      summary: 股票分时
      description: 获取指定股票的当日分时数据
      operationId: stockMinute
      parameters:
        - name: stockCode
          in: path
          required: true
          description: 股票代码，6位数字
          schema:
            type: string
            pattern: "^[0-9]{6}$"
      responses:
        "200":
          description: 股票分时数据
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/IndexMinutePoint"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /api/v1/stocks/quotes:
    post:
      tags: [股票]
      summary: 批量股票行情
      description: 批量获取股票行情数据，最多支持50个股票代码；可通过 freshness 选择平衡缓存或实时优先策略
      operationId: stockQuotes
      security:
        - csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockQuoteRequest"
      responses:
        "200":
          description: 股票行情数据
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        additionalProperties:
                          $ref: "#/components/schemas/StockQuote"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/market/stock-ranking/{type}:
    get:
      tags: [市场]
      summary: 股票排行榜
      description: 获取股票涨幅、跌幅或成交量排行榜
      operationId: stockRanking
      parameters:
        - name: type
          in: path
          required: true
          description: 排行类型，gainers为涨幅榜，losers为跌幅榜，volume为成交量榜
          schema:
            type: string
            enum: [gainers, losers, volume]
        - name: size
          in: query
          required: false
          description: 返回数量，默认10
          schema:
            type: integer
            minimum: 1
            default: 10
      responses:
        "200":
          description: 股票排行列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/StockRankingItem"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/stocks/sync:
    post:
      tags: [股票]
      summary: 同步股票数据
      description: 从数据源同步股票数据，需要管理员令牌认证
      operationId: syncStocks
      security:
        - adminToken: []
          csrfCookie: []
          csrfHeader: []
      responses:
        "200":
          description: 同步结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/StockSyncResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    adminToken:
      type: http
      scheme: bearer
      description: 管理员令牌，通过 Authorization 请求头传递 Bearer Token
    csrfCookie:
      type: apiKey
      in: cookie
      name: csrf_token
      description: 先通过任意 GET 请求获取 HttpOnly CSRF Cookie
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: 使用 GET 响应头返回的令牌，值必须与 csrf_token Cookie 一致
  schemas:
    APIResponse:
      type: object
      description: 通用API响应包装
      properties:
        code:
          type: integer
          description: 状态码，0表示成功，正数为业务错误码，-1为通用错误
          example: 0
        message:
          type: string
          description: 状态消息
          example: success
        data:
          description: 响应数据
...
    IndexMinutePoint:
      type: object
      description: 指数或股票分时数据点
      properties:
        time:
          type: string
          description: 分时时间
        price:
          type: number
          format: double
          description: 最新价格或点位
        avg_price:
          type: number
          format: double
          description: 均价
        volume:
          type: integer
          format: int64
          description: 成交量
...
    StockItem:
      type: object
      description: 股票基本信息
      properties:
        stock_code:
          type: string
          description: 股票代码
          example: "600519"
        stock_name:
          type: string
          description: 股票名称
          example: 贵州茅台
        market:
          type: string
          description: 所属市场
          example: SH
        industry:
          type: string
          description: 所属行业
        list_date:
          type: string
          description: 上市日期
        total_shares:
          type: number
          format: double
          description: 总股本
        float_shares:
          type: number
          format: double
          description: 流通股本
        current_price:
          type: number
          format: double
          description: 当前价格
        change_pct:
          type: number
          format: double
          description: 涨跌幅百分比
        volume:
          type: number
          format: double
          description: 成交量
        amount:
          type: number
          format: double
          description: 成交额
        turnover_rate:
          type: number
          format: double
          description: 换手率
        pe_ratio:
          type: number
          format: double
          description: 市盈率
        pb_ratio:
          type: number
          format: double
          description: 市净率
        total_mv:
          type: number
          format: double
          description: 总市值
        pinyin:
          type: string
          description: 拼音
        pinyin_alt:
          type: string
          description: 拼音备选
    StockSearchData:
      type: object
      description: 股票搜索结果
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/StockItem"
        total:
          type: integer
          description: 总数
        page:
          type: integer
          description: 当前页码
        size:
          type: integer
          description: 每页数量
...
    StockBasicInfo:
      type: object
      description: 股票基本信息
      properties:
        stock_code:
          type: string
          description: 股票代码
        stock_name:
          type: string
          description: 股票名称
        market:
          type: string
          description: 所属市场
        industry:
          type: string
          description: 所属行业
        list_date:
          type: string
          description: 上市日期
        total_shares:
          type: number
          format: double
          description: 总股本
        float_shares:
          type: number
          format: double
          description: 流通股本
    StockQuote:
      type: object
      description: 股票行情
      properties:
        price:
          type: number
          format: double
          description: 当前价格
        open:
          type: number
          format: double
          description: 开盘价
        high:
          type: number
          format: double
          description: 最高价
        low:
          type: number
          format: double
          description: 最低价
        prev_close:
          type: number
          format: double
          description: 昨收价
        volume:
          type: number
          format: double
          description: 成交量
        amount:
          type: number
          format: double
          description: 成交额
        turnover_rate:
          type: number
          format: double
          description: 换手率
        change_pct:
          type: number
          format: double
          description: 涨跌幅百分比
        change_amt:
          type: number
          format: double
          description: 涨跌额
        bid_price:
          type: number
          format: double
          description: 买一价
        ask_price:
          type: number
          format: double
          description: 卖一价
        quote_time:
          type: string
          description: 行情时间
    StockQuoteRequest:
      type: object
      description: 股票行情请求
      required:
        - codes
      properties:
        codes:
          type: array
          items:
            type: string
            pattern: "^[0-9]{6}$"
          minItems: 1
          maxItems: 50
          description: 股票代码列表，每个代码为6位纯数字，最多50个
        freshness:
          type: string
          enum: [balanced, realtime]
          default: balanced
          description: 行情新鲜度策略；balanced 使用较长缓存窗口，realtime 使用短 TTL 与后台刷新
    KlinePoint:
      type: object
      description: K线数据点
      properties:
        date:
          type: string
          description: 日期
        open:
          type: number
          format: double
          description: 开盘价
        close:
          type: number
          format: double
          description: 收盘价
        high:
          type: number
          format: double
          description: 最高价
        low:
          type: number
          format: double
          description: 最低价
        volume:
          type: number
          format: double
          description: 成交量
        amount:
          type: number
          format: double
          description: 成交额
    StockKlineData:
      type: object
      description: K线数据
      properties:
        period:
          type: string
          description: K线周期
        klines:
          type: array
          items:
            $ref: "#/components/schemas/KlinePoint"
          description: K线数据列表
    CapitalFlowPoint:
      type: object
      description: 资金流向数据点
      properties:
        date:
          type: string
          description: 日期
        main_inflow:
          type: number
          format: double
          description: 主力流入
        main_outflow:
          type: number
          format: double
          description: 主力流出
        retail_inflow:
          type: number
          format: double
          description: 散户流入
        retail_outflow:
          type: number
          format: double
          description: 散户流出
        net_inflow:
          type: number
          format: double
          description: 净流入
    StockCapitalFlow:
      type: object
      description: 股票资金流向
      properties:
        main_net_inflow:
          type: number
          format: double
          description: 主力净流入
        retail_net_inflow:
          type: number
          format: double
          description: 散户净流入
        flow_history:
          type: array
          items:
            $ref: "#/components/schemas/CapitalFlowPoint"
          description: 资金流向历史
    FinancialQuarter:
      type: object
      description: 季度财务数据
      properties:
        report_date:
          type: string
          description: 报告期
        revenue:
          type: number
          format: double
          description: 营业收入
        net_profit:
          type: number
          format: double
          description: 净利润
        eps:
          type: number
          format: double
          description: 每股收益
        gross_margin:
          type: number
          format: double
          description: 毛利率
        net_margin:
          type: number
          format: double
          description: 净利率
        roe:
          type: number
          format: double
          description: 净资产收益率
    StockFinancials:
      type: object
      description: 股票财务数据
      properties:
        pe_ratio:
          type: number
          format: double
          description: 市盈率
        pb_ratio:
          type: number
          format: double
          description: 市净率
        roe:
          type: number
          format: double
          description: 净资产收益率
        revenue:
          type: number
          format: double
          description: 营业收入
        net_profit:
          type: number
          format: double
          description: 净利润
        eps:
          type: number
          format: double
          description: 每股收益
        gross_margin:
          type: number
          format: double
          description: 毛利率
        net_margin:
          type: number
          format: double
          description: 净利率
        quarterly:
          type: array
          items:
            $ref: "#/components/schemas/FinancialQuarter"
          description: 季度财务数据
    ShareholderItem:
      type: object
      description: 股东信息项
      properties:
        name:
          type: string
          description: 股东名称
        ratio:
          type: number
          format: double
          description: 持股比例
        change:
          type: number
          format: double
          description: 变动比例
        share_type:
          type: string
          description: 股份类型
    StockShareholders:
      type: object
      description: 股东信息
      properties:
        top10:
          type: array
          items:
            $ref: "#/components/schemas/ShareholderItem"
          description: 前十大股东
        institution_count:
          type: integer
          description: 机构投资者数量
        institution_ratio:
          type: number
          format: double
          description: 机构持股比例
    StockDetailData:
      type: object
      description: 股票详情数据
      properties:
        basic:
          $ref: "#/components/schemas/StockBasicInfo"
        quote:
          $ref: "#/components/schemas/StockQuote"
        kline:
          $ref: "#/components/schemas/StockKlineData"
        minute_data:
          type: array
          items:
            $ref: "#/components/schemas/IndexMinutePoint"
          description: 股票分时数据
        capital_flow:
          $ref: "#/components/schemas/StockCapitalFlow"
        financials:
          $ref: "#/components/schemas/StockFinancials"
        shareholders:
          $ref: "#/components/schemas/StockShareholders"
    StockRankingItem:
      type: object
      description: 股票排行项
      properties:
        rank:
          type: integer
          description: 排名
        stock_code:
          type: string
          description: 股票代码
        stock_name:
          type: string
          description: 股票名称
        change_pct:
          type: number
          format: double
          description: 涨跌幅百分比
        current_price:
          type: number
          format: double
          description: 当前价格
        volume:
          type: number
          format: double
          description: 成交量
        amount:
          type: number
          format: double
          description: 成交额
        data_source:
          type: string
          description: 行情数据来源
        update_time:
          type: string
          description: 更新时间
    StockSyncResult:
      type: object
      description: 股票同步结果
      properties:
        total:
          type: integer
          description: 总数量
        imported:
          type: integer
          description: 导入数量
        errors:
          type: integer
          description: 错误数量
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            code: -1
            message: "查询参数格式错误"
            data: null
    Unauthorized:
      description: 未授权，管理员令牌无效或缺失
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            code: -1
            message: "未提供管理员令牌"
            data: null
    Forbidden:
      description: CSRF Cookie 或请求头缺失、不一致
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            code: -1
            message: "CSRF token 验证失败"
            data: null
...
    InternalServerError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            code: -1
            message: "服务器内部错误"
            data: null
    ServiceUnavailable:
      description: 服务暂时不可用
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            code: -1
            message: "服务暂时不可用"
            data: null